Add Callouts preview accordion on SubspacesList default subspace template selector
