+++ src/domain/journey/space/pages/SpaceSubspaces/AdminChallengesPage.graphql
@@ -27,6 +27,23 @@ query AdminSpaceSubspacesPage($spaceId: UUID_NAMEID!) {
           }
           collaboration {
             id
+            callouts {
+              id
+              type
+              sortOrder
+              framing  {
+                id
+                profile {
+                  id
+                  displayName
+                  description
+                  flowStateTagset: tagset(tagsetName: FLOW_STATE) {
+                    id
+                    tags
+                  }
+                }
+              }
+            }
             innovationFlow {
               id
               profile {
